talk page - controller rework
- allows talk page to be created via new discussion post
- grouped revID & talk page calls where possible
- progress indicator on initial fetch

+++ WikipediaUnitTests/Code/TalkPageTestHelpers.swift
@@ -75,7 +75,7 @@ class TalkPageTestHelpers {
         }
     }
     
-    static func networkTalkPage(for urlString: String, jsonType: TalkPageJSONType = .original, revisionId: Int64) -> NetworkTalkPage? {
+    static func networkTalkPage(for urlString: String, jsonType: TalkPageJSONType = .original, revisionId: Int) -> NetworkTalkPage? {
         let session = Session.shared
         do {
             let result: NetworkBase = try session.jsonDecodeData(data: jsonType.json)
